+++ plugins/CoreHome/CoreHome.php
@@ -30,32 +30,12 @@ class CoreHome extends \Piwik\Plugin
             'AssetManager.getJavaScriptFiles'        => 'getJsFiles',
             'AssetManager.filterMergedJavaScripts'   => 'filterMergedJavaScripts',
             'Translate.getClientSideTranslationKeys' => 'getClientSideTranslationKeys',
-            'Live.getAllVisitorDetails'              => 'extendVisitorDetails',
         );
     }
 
     public function filterMergedJavaScripts(&$mergedContent)
     {
         $mergedContent = preg_replace('/(sourceMappingURL=(.*?).map)/', '', $mergedContent);
-    }
-
-    public function extendVisitorDetails(&$visitor, $details)
-    {
-        $instance = new Visitor($details);
-
-        $visitor['userId']                      = $instance->getUserId();
-        $visitor['visitorType']                 = $instance->getVisitorReturning();
-        $visitor['visitorTypeIcon']             = $instance->getVisitorReturningIcon();
-        $visitor['visitConverted']              = $instance->isVisitorGoalConverted();
-        $visitor['visitConvertedIcon']          = $instance->getVisitorGoalConvertedIcon();
-        $visitor['visitCount']                  = $instance->getVisitCount();
-        $visitor['firstActionTimestamp']        = $instance->getTimestampFirstAction();
-        $visitor['visitEcommerceStatus']        = $instance->getVisitEcommerceStatus();
-        $visitor['visitEcommerceStatusIcon']    = $instance->getVisitEcommerceStatusIcon();
-        $visitor['daysSinceFirstVisit']         = $instance->getDaysSinceFirstVisit();
-        $visitor['daysSinceLastEcommerceOrder'] = $instance->getDaysSinceLastEcommerceOrder();
-        $visitor['visitDuration']               = $instance->getVisitLength();
-        $visitor['visitDurationPretty']         = $instance->getVisitLengthPretty();
     }
 
     public function getStylesheetFiles(&$stylesheets)