+++ plugins/CustomVariables/CustomVariables.php
@@ -20,10 +20,10 @@ class CustomVariables extends \Piwik\Plugin
     public function registerEvents()
     {
         return array(
-            'Live.getAllVisitorDetails'       => 'extendVisitorDetails',
             'AssetManager.getJavaScriptFiles' => 'getJsFiles',
             'Translate.getClientSideTranslationKeys' => 'getClientSideTranslationKeys',
             'AssetManager.getStylesheetFiles'  => 'getStylesheetFiles',
+            'Actions.getCustomActionDimensionFieldsAndJoins' => 'provideActionDimensionFields'
         );
     }
 
@@ -35,24 +35,6 @@ class CustomVariables extends \Piwik\Plugin
     public function uninstall()
     {
         Model::uninstall();
-    }
-
-    public function extendVisitorDetails(&$visitor, $details)
-    {
-        $customVariables = array();
-
-        $maxCustomVariables = self::getNumUsableCustomVariables();
-
-        for ($i = 1; $i <= $maxCustomVariables; $i++) {
-            if (!empty($details['custom_var_k' . $i])) {
-                $customVariables[$i] = array(
-                    'customVariableName' .  $i => $details['custom_var_k' . $i],
-                    'customVariableValue' . $i => $details['custom_var_v' . $i],
-                );
-            }
-        }
-
-        $visitor['customVariables'] = $customVariables;
     }
 
     /**
@@ -136,4 +118,13 @@ class CustomVariables extends \Piwik\Plugin
         $jsFiles[] = "plugins/CustomVariables/angularjs/manage-custom-vars/manage-custom-vars.directive.js";
     }
 
+    public function provideActionDimensionFields(&$fields, &$joins)
+    {
+        $maxCustomVariables = CustomVariables::getNumUsableCustomVariables();
+
+        for ($i = 1; $i <= $maxCustomVariables; $i++) {
+            $fields[] = 'custom_var_k' . $i;
+            $fields[] = 'custom_var_v' . $i;
+        }
+    }
 }